feat: refactor package structure
