+++ lib/features/guide_detail_view/data/models/guide_details.dart
