+++ lib/features/guide_view/data/models/guide_data.dart
