+++ lib/features/home_view/keep_alive_home_view_providers.dart
@@ -6,7 +6,7 @@ import "../academic_calendar/repository/academic_calendar_repo.dart";
 import "../app_streak/business/get_days_use_case.dart";
 import "../buildings_view/repository/buildings_repository.dart";
 import "../departments/departments_view/repository/departments_repository.dart";
-import "../guide_view/repository/guide_repository.dart";
+import "../guide_view/data/repository/guide_repository.dart";
 import "../parkings/parkings_view/repository/parkings_repository.dart";
 import "../science_club/science_clubs_view/repository/science_clubs_repository.dart";
 import "home_view.dart";
