+++ lib/features/home_view/widgets/news_section.dart
@@ -9,8 +9,8 @@ import "../../../utils/context_extensions.dart";
 import "../../../widgets/big_preview_card.dart";
 import "../../../widgets/my_error_widget.dart";
 import "../../../widgets/subsection_header.dart";
-import "../../guide_view/models/guide_data.dart";
-import "../../guide_view/repository/guide_repository.dart";
+import "../../guide_view/data/models/guide_data.dart";
+import "../../guide_view/data/repository/guide_repository.dart";
 import "../../navigator/utils/navigation_commands.dart";
 import "loading_widgets/big_scrollable_section_loading.dart";
 import "paddings.dart";
